+++ Data/ExportedCategoriesRetail.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-10-16 14-17-24 ]] --
+-- [[ Exported at 2022-10-24 20-43-40 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --