+++ Localization/deDE.ExtraWrathStrings.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-10-16 10-12-17 ]] --
+-- [[ Exported at 2022-10-24 19-19-51 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --